fix: memo layout
